Move mapping profile validation to service layer
